requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch>=1.8.0
torchvision>=0.9.0
ultralytics>=8.0.0
numpy>=1.23.0
Pillow>=7.1.2
pydantic>=2.0.0
python-dotenv>=0.19.0
